Refactor #76 - clear packages of compose
