+++ app/src/main/java/com/nlab/reminder/core/android/designsystem/theme/Theme.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.domain.common.android.designsystem.theme
+package com.nlab.reminder.core.android.designsystem.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
