+++ app/src/main/java/com/nlab/reminder/core/android/designsystem/theme/Type.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.domain.common.android.designsystem.theme
+package com.nlab.reminder.core.android.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
