+++ app/src/main/java/com/nlab/reminder/domain/common/tag/view/TagRenameDialog.kt
@@ -63,9 +63,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.nlab.reminder.R
-import com.nlab.reminder.domain.common.android.designsystem.component.ColorPressButton
-import com.nlab.reminder.domain.common.android.designsystem.component.ReminderThemeDialog
-import com.nlab.reminder.domain.common.android.designsystem.theme.ReminderTheme
+import com.nlab.reminder.core.android.designsystem.component.ColorPressButton
+import com.nlab.reminder.core.android.designsystem.component.ReminderThemeDialog
+import com.nlab.reminder.core.android.designsystem.theme.ReminderTheme
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.debounce
