+++ app/src/main/java/com/nlab/reminder/domain/feature/home/view/BottomContainer.kt
@@ -36,7 +36,7 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.nlab.reminder.domain.common.android.designsystem.theme.ReminderTheme
+import com.nlab.reminder.core.android.designsystem.theme.ReminderTheme
 
 /**
  * @author Doohyun
